Added Examples and Demo
Added 3 sample cards. Changed the name of changeMaterialWithLoc to changeMatWithLoc to shorten it. Added a short demo code for the use of changeMathWithLoc
