+++ Brain Games/Assets/Prefabs/CardMain.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64234a8ec35ccb14687fd040562354db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
