+++ Brain Games/Library/LastSceneManagerSetup.txt
@@ -1,1 +1,4 @@
-sceneSetups: []
+sceneSetups:
+- path: 
+  isLoaded: 1
+  isActive: 1
